requirements.txt
sentence-transformers==2.2.2
huggingface_hub
langchain==0.0.225
flask
pypdf
python-dotenv
pinecone-client
-e .